+++ src/main/java/com/spotify/sparkey/cleaner/MappedByteBufferCleaner.java
@@ -1,0 +1,7 @@
+package com.spotify.sparkey.cleaner;
+
+import java.nio.MappedByteBuffer;
+
+public interface MappedByteBufferCleaner {
+    void cleanup(MappedByteBuffer... chunks);
+}
